[refac]: moved class ODDFactory to own file
