+++ test/conftest.py
@@ -15,9 +15,9 @@ from ssrq_cli.xml_utils import ext_etree
 
 from utils.constants import SPECIFIED_ELEMENTS, SRC_DIR, XSLTS
 from utils.main import (
-    ODDFactory,
     load_config,
 )
+from utils.ODDFactory import ODDFactory
 from utils.SSRQSchema import SSRQSchema
 from utils.SSRQSchemaType import SSRQSchemaType
